fix(components/date_picker): REFACTOR callbacks
