+++ DashMonitor/app/views/components/__init__.py
@@ -6,6 +6,6 @@ from DashMonitor.app.views.components.header import Header, HeaderTabBtn
 from DashMonitor.app.views.components.main import Main, MainTabPanel
 from DashMonitor.app.views.components.footer import Footer
 from DashMonitor.app.views.components.card import Card
-from DashMonitor.app.views.components.date_picker import DatePicker
+from DashMonitor.app.views.components.date_picker.date_picker import DatePicker
 from DashMonitor.app.views.components.table.table import Table
 from DashMonitor.app.views.components.gauge_chart import GaugeChart
